+++ CHANGELOG.md
@@ -48,3 +48,4 @@
    - FEATURE: support for choice/case statements (initial support with no validation)
    - FEATURE: support UINT64/INT64 values.
    - FEATURE: raise an exception `Cannot assign a value to ......` when trying to assign values to containing nodes.
+   - FEATURE: add `session.from_xmlstr` to import data without jinja2 processing from disk.